chore: move `export_nep_330_build_command` to `--no-docker` build flow
